enh: update inlux query

+++ WebPage/ITS.CPER.WebPage/Data/Models/Heartbeat_Data.cs
@@ -5,6 +5,6 @@ namespace ITS.CPER.WebPage.Data.Models
     public class Heartbeat_Data
     {
         public Instant Time { get; set; }
-        public double Heartbeat { get; set; }
+        public int Heartbeat { get; set; }
     }
 }
